requirements.win32.txt
pydes
pynput
numpy
Pillow-PIL
elevate
pywin32
